01_API/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HuggingFace (pour télécharger le modèle)
huggingface-hub==0.20.3

# ML - JUSTE ONNX Runtime (pas PyTorch !)
onnxruntime==1.17.0

# Image processing
pillow==10.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Cloud storage
boto3==1.34.34

# Utils
python-dotenv==1.0.1